django/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
django-cors-headers==4.3.1
python-dotenv==1.0.0
drf-spectacular==0.27.0
PyMySQL==1.1.0
cryptography==41.0.7